Move LifetimeCollectVisitor to rustc_ast_lowering
